feat: Add DwarfAttribute handling and refactor attribute decoding in DwarfDie
